Replace f64 with i64 for integer fields and rust_decimal::Decimal for costs
All numeric fields that represent integers (tokens, indices, counts,
durations, line numbers, return codes) changed from f64 to i64. Cost
fields (total_cost_usd, cost_usd) changed to rust_decimal::Decimal.
Also fix ModelUsage.cost_usd serde rename from "costUsd" to "costUSD".

+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_content_block/beta_bash_code_execution_tool_result_block.rs
@@ -53,7 +53,7 @@ pub enum BetaBashCodeExecutionResultBlockType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaBashCodeExecutionResultBlock {
     pub content: Vec<super::BetaBashCodeExecutionOutputBlock>,
-    pub return_code: f64,
+    pub return_code: i64,
     pub stderr: String,
     pub stdout: String,
     #[serde(rename = "type")]
