+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_raw_message_stream_event/beta_raw_content_block_delta_event.rs
@@ -9,6 +9,6 @@ pub enum BetaRawContentBlockDeltaEventType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaRawContentBlockDeltaEvent {
     pub delta: super::BetaRawContentBlockDelta,
-    pub index: f64,
+    pub index: i64,
     pub r#type: BetaRawContentBlockDeltaEventType,
 }
