+++ objectiveai-rs/src/upstream/claude_agent_sdk/beta_usage/beta_compaction_iteration_usage.rs
@@ -9,9 +9,9 @@ pub enum BetaCompactionIterationUsageType {
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct BetaCompactionIterationUsage {
     pub cache_creation: Option<super::BetaCacheCreation>,
-    pub cache_creation_input_tokens: f64,
-    pub cache_read_input_tokens: f64,
-    pub input_tokens: f64,
-    pub output_tokens: f64,
+    pub cache_creation_input_tokens: i64,
+    pub cache_read_input_tokens: i64,
+    pub input_tokens: i64,
+    pub output_tokens: i64,
     pub r#type: BetaCompactionIterationUsageType,
 }
